Fixed tests Moved test_IntegrityReferentWorkflow to correct file (testCustomWorkflows.py) Give also DG_GROUP_ID manage item when item is "returned_to_proposing_group_proposed_to_dg"
