+++ src/Products/MeetingLalouviere/tests/testFaceted.py
@@ -30,7 +30,7 @@ class testFaceted(MeetingLalouviereTestCase, mctf):
     """Tests the faceted navigation."""
 
     def _orgs_to_exclude_from_filter(self):
-        return (self.direction_generale_uid,)
+        return (self.direction_generale_validation_uid,)
 
 
 def test_suite():
