+++ src/Products/MeetingLalouviere/tests/testSearches.py
@@ -2,24 +2,7 @@
 #
 # File: testMeetingConfig.py
 #
-# Copyright (c) 2015 by Imio.be
-#
 # GNU General Public License (GPL)
-#
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License
-# as published by the Free Software Foundation; either version 2
-# of the License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
-# 02110-1301, USA.
 #
 
 from collective.compoundcriterion.interfaces import ICompoundCriterionFilter
@@ -37,7 +20,7 @@ class testSearches(MeetingLalouviereTestCase, mcts):
 
     def setUp(self):
         super(testSearches, self).setUp()
-        self._removePrincipalFromGroups("pmManager", [self.direction_generale_directors])
+        self._removePrincipalFromGroups("pmManager", [self.direction_generale_validation_directors])
 
     def test_pm_SearchItemsToValidateOfHighestHierarchicLevelReturnsEveryLevels(self):
         pass
